Refactor scanner into modules and add detection flag
